Reorganize utils module to sentinel package
This commit:
1. Moves utils implementation to sentinel/utils package
2. Creates backward compatibility layers for existing imports
3. Implements adaptive plasticity module and metrics
4. Adds unit tests for utils structure
5. Sets up stubs for additional utils components
