+++ sentinel/utils/pruning/fixed_pruning_module.py
@@ -1,0 +1,51 @@
+"""
+Fixed Pruning Module for transformer models.
+
+This module provides a simplified interface for attention head pruning with
+fixed pruning decisions (rather than dynamic ones during training).
+"""
+
+class FixedPruningModule:
+    """
+    Pruning module that provides fixed pruning functionality.
+    This is a placeholder that will be implemented when pruning is moved.
+    """
+    
+    def __init__(self, model_name):
+        """Initialize with model name."""
+        self.model_name = model_name
+        # These will be set when the model is loaded
+        self.num_layers = 0
+        self.num_heads = 0
+        self.model = None
+    
+    def load_model(self):
+        """Load the model (placeholder)."""
+        # This is a placeholder - would load the actual model
+        print(f"[Placeholder] Loading model {self.model_name}")
+        self.num_layers = 12  # Placeholder value
+        self.num_heads = 12   # Placeholder value
+        
+        class DummyModel:
+            def __init__(self):
+                self.params = {}
+        
+        self.model = DummyModel()
+        return True
+    
+    def prune_head(self, params, layer_idx, head_idx):
+        """Prune a specific head (placeholder)."""
+        # This is a placeholder - would actually modify params
+        print(f"[Placeholder] Pruning head at layer {layer_idx}, head {head_idx}")
+        return params.copy()
+    
+    def evaluate_perplexity(self, params, text):
+        """Evaluate perplexity (placeholder)."""
+        # This is a placeholder - would compute actual perplexity
+        import random
+        return random.uniform(5.0, 20.0)
+    
+    def generate_text(self, params, prompt, max_length=100):
+        """Generate text (placeholder)."""
+        # This is a placeholder - would generate actual text
+        return prompt + " [This is placeholder generated text]"
