+++ sentinel/utils/pruning/growth.py
@@ -1,0 +1,138 @@
+"""
+Growth strategies for attention heads.
+
+This module provides functionality to grow new attention heads
+after pruning, using different strategies for head selection.
+"""
+
+from typing import List, Tuple, Callable, Any
+
+
+def determine_active_heads(pruning_module, params):
+    """
+    Determine which heads are currently active in the model.
+    
+    Args:
+        pruning_module: The pruning module instance
+        params: Model parameters
+        
+    Returns:
+        List of (layer_idx, head_idx) tuples for active heads
+    """
+    # Placeholder implementation
+    active_heads = []
+    for layer_idx in range(pruning_module.num_layers):
+        for head_idx in range(pruning_module.num_heads):
+            # In a real implementation, would check params to see if head is active
+            import random
+            # For placeholder, include 80% of heads as active
+            if random.random() < 0.8:
+                active_heads.append((layer_idx, head_idx))
+    
+    return active_heads
+
+
+def grow_attention_heads_gradually(
+    pruning_module,
+    params: Any,
+    active_heads: List[Tuple[int, int]],
+    growth_percentage: float = 0.1,
+    strategy: str = "gradient_sensitivity",
+    initial_scale: float = 0.01,
+    warmup_steps: int = 100
+) -> Tuple[Any, int, List[Tuple[int, int]], Callable]:
+    """
+    Grow new attention heads gradually.
+    
+    Args:
+        pruning_module: The pruning module instance
+        params: Model parameters
+        active_heads: List of currently active heads
+        growth_percentage: Percentage of total heads to grow
+        strategy: Growth strategy name
+        initial_scale: Initial scale for new heads
+        warmup_steps: Steps to gradually warm up new heads
+        
+    Returns:
+        Tuple of:
+        - Updated parameters
+        - Number of heads added
+        - List of added head indices
+        - Warmup schedule function
+    """
+    # Placeholder implementation
+    
+    # Calculate how many heads to add
+    total_heads = pruning_module.num_layers * pruning_module.num_heads
+    heads_to_add = max(1, int(total_heads * growth_percentage))
+    
+    # Choose which heads to add based on strategy
+    added_heads = []
+    
+    # For placeholder, just add random heads that aren't already active
+    import random
+    all_possible_heads = [
+        (layer_idx, head_idx) 
+        for layer_idx in range(pruning_module.num_layers) 
+        for head_idx in range(pruning_module.num_heads)
+    ]
+    inactive_heads = [head for head in all_possible_heads if head not in active_heads]
+    
+    if len(inactive_heads) < heads_to_add:
+        heads_to_add = len(inactive_heads)
+    
+    if heads_to_add > 0:
+        added_heads = random.sample(inactive_heads, heads_to_add)
+    
+    # Copy parameters for modification
+    grown_params = params.copy()
+    
+    # In a real implementation, would actually modify params here
+    
+    # Define warmup schedule
+    def warmup_schedule(step):
+        """Return the scaling factor based on current step."""
+        if step >= warmup_steps:
+            return 1.0
+        else:
+            return initial_scale + (1.0 - initial_scale) * (step / warmup_steps)
+    
+    return grown_params, len(added_heads), added_heads, warmup_schedule
+
+
+def get_strategy(strategy_name, pruning_module):
+    """
+    Get the requested growth strategy.
+    
+    Args:
+        strategy_name: Name of the strategy
+        pruning_module: The pruning module instance
+        
+    Returns:
+        Growth strategy object
+    """
+    # Placeholder implementation
+    class DummyStrategy:
+        def __init__(self, pruning_module):
+            self.pruning_module = pruning_module
+        
+        def select_heads_to_grow(self, params, active_heads, num_to_grow):
+            """Select heads to grow."""
+            # In a real implementation, would use a strategy-specific algorithm
+            import random
+            all_possible_heads = [
+                (layer_idx, head_idx) 
+                for layer_idx in range(self.pruning_module.num_layers) 
+                for head_idx in range(self.pruning_module.num_heads)
+            ]
+            inactive_heads = [head for head in all_possible_heads if head not in active_heads]
+            
+            if len(inactive_heads) < num_to_grow:
+                num_to_grow = len(inactive_heads)
+            
+            if num_to_grow > 0:
+                return random.sample(inactive_heads, num_to_grow)
+            else:
+                return []
+    
+    return DummyStrategy(pruning_module)
