+++ sentinel/utils/pruning/strategies.py
@@ -1,0 +1,101 @@
+"""
+Pruning strategies for attention heads.
+
+This module provides different strategies for determining head importance
+and making pruning decisions.
+"""
+
+class PruningStrategy:
+    """Base class for pruning strategies."""
+    
+    def __init__(self, pruning_module):
+        """Initialize with the pruning module."""
+        self.pruning_module = pruning_module
+    
+    def get_head_importance(self, params):
+        """
+        Calculate importance scores for each head.
+        
+        Args:
+            params: Model parameters
+            
+        Returns:
+            List of (layer_idx, head_idx, importance) tuples
+        """
+        raise NotImplementedError("Subclasses must implement this method")
+
+
+class EntropyStrategy(PruningStrategy):
+    """Prune heads based on attention entropy."""
+    
+    def get_head_importance(self, params):
+        """Higher entropy = lower importance."""
+        # Placeholder implementation
+        import random
+        
+        importance_list = []
+        for layer_idx in range(self.pruning_module.num_layers):
+            for head_idx in range(self.pruning_module.num_heads):
+                # Random importance score for placeholder
+                importance = random.uniform(0.0, 1.0)
+                importance_list.append((layer_idx, head_idx, importance))
+        
+        return importance_list
+
+
+class MagnitudeStrategy(PruningStrategy):
+    """Prune heads based on weight magnitudes."""
+    
+    def get_head_importance(self, params):
+        """Lower magnitude = lower importance."""
+        # Placeholder implementation
+        import random
+        
+        importance_list = []
+        for layer_idx in range(self.pruning_module.num_layers):
+            for head_idx in range(self.pruning_module.num_heads):
+                # Random importance score for placeholder
+                importance = random.uniform(0.0, 1.0)
+                importance_list.append((layer_idx, head_idx, importance))
+        
+        return importance_list
+
+
+class RandomStrategy(PruningStrategy):
+    """Randomly prune heads (baseline)."""
+    
+    def get_head_importance(self, params):
+        """Random importance scores."""
+        import random
+        
+        importance_list = []
+        for layer_idx in range(self.pruning_module.num_layers):
+            for head_idx in range(self.pruning_module.num_heads):
+                # Random importance score
+                importance = random.random()
+                importance_list.append((layer_idx, head_idx, importance))
+        
+        return importance_list
+
+
+def get_strategy(strategy_name, pruning_module):
+    """
+    Get the requested pruning strategy.
+    
+    Args:
+        strategy_name: Name of the strategy ("entropy", "magnitude", "random")
+        pruning_module: The pruning module instance
+        
+    Returns:
+        PruningStrategy instance
+    """
+    strategies = {
+        "entropy": EntropyStrategy,
+        "magnitude": MagnitudeStrategy,
+        "random": RandomStrategy
+    }
+    
+    if strategy_name not in strategies:
+        raise ValueError(f"Unknown pruning strategy: {strategy_name}")
+    
+    return strategies[strategy_name](pruning_module)
