+++ sentinel/utils/train_utils.py
@@ -1,0 +1,142 @@
+"""
+Training utilities for transformer models.
+
+This module provides utilities for fine-tuning models after pruning
+or growing new attention heads.
+"""
+
+import torch
+from typing import Dict, Any, Optional, List, Tuple, Callable
+
+
+class FineTuner:
+    """
+    Utility for fine-tuning transformer models after pruning or growth.
+    """
+    
+    def __init__(
+        self, 
+        pruning_module,
+        dataset,
+        learning_rate: float = 1e-4,
+        head_lr_manager: Optional[Any] = None,
+        batch_size: int = 8,
+        warmup_steps: int = 100,
+        weight_decay: float = 0.01
+    ):
+        """
+        Initialize the fine-tuner.
+        
+        Args:
+            pruning_module: The pruning module instance
+            dataset: Dataset to use for fine-tuning
+            learning_rate: Base learning rate
+            head_lr_manager: Optional manager for head-specific learning rates
+            batch_size: Batch size for training
+            warmup_steps: Learning rate warmup steps
+            weight_decay: Weight decay for regularization
+        """
+        self.pruning_module = pruning_module
+        self.dataset = dataset
+        self.learning_rate = learning_rate
+        self.head_lr_manager = head_lr_manager
+        self.batch_size = batch_size
+        self.warmup_steps = warmup_steps
+        self.weight_decay = weight_decay
+        
+        # Training state
+        self.step = 0
+        self.params = None
+        self.optimizer = None
+        
+        # Set up initial parameters from the pruning module
+        if hasattr(pruning_module, "model") and hasattr(pruning_module.model, "params"):
+            self.params = pruning_module.model.params.copy()
+    
+    def set_params(self, params: Dict[str, Any]) -> None:
+        """
+        Set the model parameters for fine-tuning.
+        
+        Args:
+            params: Model parameters
+        """
+        self.params = params.copy()
+    
+    def get_params(self) -> Dict[str, Any]:
+        """
+        Get the current model parameters.
+        
+        Returns:
+            Current model parameters
+        """
+        return self.params.copy()
+    
+    def train_step(self) -> float:
+        """
+        Perform a single training step.
+        
+        Returns:
+            Training loss for this step
+        """
+        # This is a placeholder implementation
+        # In a real implementation, would:
+        # 1. Get a batch from the dataset
+        # 2. Update learning rates if needed
+        # 3. Do forward and backward passes
+        # 4. Update model parameters
+        
+        import random
+        
+        # Simulate training with random loss improvement
+        loss = 10.0 * (0.99 ** self.step) + random.uniform(-0.5, 0.5)
+        
+        # Increment step counter
+        self.step += 1
+        
+        return loss
+    
+    def train(self, num_steps: int) -> Tuple[Dict[str, Any], Dict[str, Any]]:
+        """
+        Train for the specified number of steps.
+        
+        Args:
+            num_steps: Number of training steps
+            
+        Returns:
+            Tuple of (final_params, training_metrics)
+        """
+        training_losses = []
+        
+        for _ in range(num_steps):
+            loss = self.train_step()
+            training_losses.append(loss)
+        
+        # Return final parameters and metrics
+        return self.get_params(), {
+            "train_losses": training_losses,
+            "final_loss": training_losses[-1] if training_losses else None
+        }
+    
+    def evaluate(self, text_samples: List[str]) -> Dict[str, Any]:
+        """
+        Evaluate the model on text samples.
+        
+        Args:
+            text_samples: List of text samples to evaluate
+            
+        Returns:
+            Dictionary with evaluation metrics
+        """
+        # This is a placeholder implementation
+        # In a real implementation, would compute perplexity and other metrics
+        
+        import random
+        
+        perplexities = [
+            random.uniform(5.0, 20.0) * (0.95 ** self.step) for _ in text_samples
+        ]
+        
+        return {
+            "perplexities": perplexities,
+            "average_perplexity": sum(perplexities) / len(perplexities) if perplexities else 0.0
+        }
