+++ tests/__init__.py
@@ -1,1 +1,1 @@
-# Tests package
+# Tests package for Sentinel-AI
