+++ tests/unit/__init__.py
@@ -1,1 +1,1 @@
-# Unit tests package
+# Unit tests package for Sentinel-AI
