+++ tests/unit/utils/__init__.py
@@ -1,0 +1,1 @@
+# Unit tests for utils module
